#109 - lib: core sources are a mess
