+++ src/lib/core/qquote.rs
@@ -5,12 +5,12 @@
 use crate::{
     core::{
         exception::{self, Condition, Exception},
-        lib::Core as _,
         mu::Mu,
         namespace::Namespace,
         reader::{Core as _, Reader},
         types::{Tag, Type},
     },
+    streams::read::Core as _,
     types::{
         cons::{Cons, Core as _},
         stream::{Core as _, Stream},
