+++ src/lib/core/reader.rs
@@ -58,8 +58,8 @@ impl Core for Reader {
     fn build(&self, mu: &Mu) -> Self {
         Reader {
             append: Namespace::intern_symbol(mu, mu.lib_ns, "append".to_string(), Tag::nil()),
-            eol: self.eol,
             cons: Namespace::intern_symbol(mu, mu.lib_ns, "cons".to_string(), Tag::nil()),
+            eol: self.eol,
         }
     }
