+++ src/lib/features/nix/nix_.rs
@@ -5,9 +5,9 @@
 #![allow(unused_imports)]
 use crate::{
     core::{
-        apply::CoreFunctionDef,
         exception::{self, Condition, Exception},
         frame::Frame,
+        lib::CoreFunctionDef,
         mu::Mu,
         types::Tag,
     },
