+++ src/lib/streams/write.rs
@@ -1,0 +1,110 @@
+//  SPDX-FileCopyrightText: Copyright 2022 James M. Putnam (putnamjm.design@gmail.com)
+//  SPDX-License-Identifier: MIT
+
+//! stream write functions
+#[allow(unused_imports)]
+use crate::{
+    async_::context::{Context, Core as _, MuFunction as _},
+    core::{
+        apply::{Core as _, MuFunction as _},
+        compile::{Compile, MuFunction as _},
+        dynamic::MuFunction as _,
+        exception::{self, Condition, Exception, MuFunction as _},
+        frame::{Frame, MuFunction as _},
+        gc::{Gc, MuFunction as _},
+        heap::{Heap, MuFunction as _},
+        mu::Mu,
+        namespace::{MuFunction as _, Namespace},
+        qquote::QqReader,
+        reader::{Core as _, Reader},
+        readtable::{map_char_syntax, SyntaxType},
+        types::{MuFunction as _, Tag, Type},
+        utime::MuFunction as _,
+    },
+    types::{
+        char::{Char, Core as _},
+        cons::{Cons, Core as _, MuFunction as _},
+        fixnum::{Core as _, Fixnum, MuFunction as _},
+        float::{Core as _, Float, MuFunction as _},
+        function::{Core as _, Function},
+        stream::{Core as _, Stream},
+        streams::MuFunction as _,
+        struct_::{Core as _, MuFunction as _, Struct},
+        symbol::{Core as _, MuFunction as _, Symbol, UNBOUND},
+        vector::{Core as _, MuFunction as _, Vector},
+    },
+};
+
+pub trait Core {
+    fn write_stream(&self, _: Tag, _: bool, _: Tag) -> exception::Result<()>;
+    fn write_string(&self, _: &str, _: Tag) -> exception::Result<()>;
+}
+
+impl Core for Mu {
+    fn write_stream(&self, tag: Tag, escape: bool, stream: Tag) -> exception::Result<()> {
+        if stream.type_of() != Type::Stream {
+            panic!("{:?}", stream.type_of())
+        }
+
+        match tag.type_of() {
+            Type::AsyncId => Context::write(self, tag, escape, stream),
+            Type::Char => Char::write(self, tag, escape, stream),
+            Type::Cons => Cons::write(self, tag, escape, stream),
+            Type::Fixnum => Fixnum::write(self, tag, escape, stream),
+            Type::Float => Float::write(self, tag, escape, stream),
+            Type::Function => Function::write(self, tag, escape, stream),
+            Type::Keyword => Symbol::write(self, tag, escape, stream),
+            Type::Null => Symbol::write(self, tag, escape, stream),
+            Type::Stream => Stream::write(self, tag, escape, stream),
+            Type::Struct => Struct::write(self, tag, escape, stream),
+            Type::Symbol => Symbol::write(self, tag, escape, stream),
+            Type::Vector => Vector::write(self, tag, escape, stream),
+            _ => panic!(),
+        }
+    }
+
+    fn write_string(&self, str: &str, stream: Tag) -> exception::Result<()> {
+        if stream.type_of() != Type::Stream {
+            panic!("{:?}", stream.type_of())
+        }
+
+        for ch in str.chars() {
+            match Stream::write_char(self, stream, ch) {
+                Ok(_) => (),
+                Err(e) => return Err(e),
+            }
+        }
+
+        Ok(())
+    }
+}
+
+pub trait MuFunction {
+    fn lib_write(_: &Mu, _: &mut Frame) -> exception::Result<()>;
+}
+
+impl MuFunction for Mu {
+    fn lib_write(mu: &Mu, fp: &mut Frame) -> exception::Result<()> {
+        let value = fp.argv[0];
+        let escape = fp.argv[1];
+        let stream = fp.argv[2];
+
+        fp.value = match mu.fp_argv_check("write", &[Type::T, Type::T, Type::Stream], fp) {
+            Ok(_) => match mu.write_stream(value, !escape.null_(), stream) {
+                Ok(_) => value,
+                Err(e) => return Err(e),
+            },
+            Err(e) => return Err(e),
+        };
+
+        Ok(())
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    #[test]
+    fn it_works() {
+        assert_eq!(2 + 2, 4);
+    }
+}
