+++ src/lib/types/function.rs
@@ -7,10 +7,10 @@ use crate::{
         exception,
         gc::Core as _,
         indirect::IndirectTag,
-        lib::Core as _,
         mu::Mu,
         types::{Tag, TagType, Type},
     },
+    streams::write::Core as _,
     types::{
         fixnum::Fixnum,
         symbol::Symbol,
